Finished the `create_df` function.
This function will return a DataFrame that is ready to be saved as a .csv file.

NOTE: For the future it may be possible to speed up this function by replacing the for loop in the function with a pandas (possibly) vectorized operation. Although do note is it much faster than the previous implementation.